+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-5929 Changed file listing API to Readdir+
 * VFS-6176 Fixed broken Oneprovider dropdown on global map
 * VFS-6115 Service Pack 1 for embedded Oneprovider GUI
 * VFS-5767 Added metadata modal